ARM: dts: msm: Introduce dummy demura panel ID node
ABL expects a demura panel ID node to always be present under
the DSI display definition. If this node is not present, ABL
is unable to insert it and demura continuous splash can not
be brought online.

Introduce a dummy demura panel ID node into the device tree.
This node is given a panel ID of 0 to allow the kernel to
specially handle it.

Change-Id: I4da25f02152c6410d4e67868f702149a7a81bd4c